pyproject.toml
[tool.poetry]
name = "aci-dns-check"
version = "1.0.0"
description = "EaaS extension to monitor DNS resolution health in ACI"
authors = ["CloudBlue"]
license = "MIT"
packages = [
    { include = "aci_dns_check" },
]

[tool.poetry.plugins."connect.eaas.ext"]
"eventsapp" = "aci_dns_check.extension:AciDnsCheckExt"

[tool.poetry.dependencies]
python = ">=3.9,<4"
connect-eaas-core = ">=33.0,<38"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
